Fix and update device docs, datum sequence, and images (#748)
* fix docs datum + imgs

* Add tools > touchscreen description in docs

* added more coverage
