+++ simulator/sequences/tools-check-sd.txt
@@ -6,6 +6,12 @@ press BUTTON_A
 
 screenshot tools-options.png
 
+# Device tests
+press BUTTON_B
+press BUTTON_A
+
+screenshot device-tests-options.png
+
 # Check SD Card
 press BUTTON_A
