+++ src/krux/pages/file_manager.py
@@ -45,6 +45,7 @@ class FileManager(Page):
         import os
 
         path = SD_ROOT_PATH
+        status = ""
         while True:
             # if is a dir then list all files in it
             if SDHandler.dir_exists(path):
