+++ tests/pages/test_device_tests.py
@@ -127,6 +127,51 @@ def mock_zlib_code(mocker):
     )
 
 
+def test_sd_check_no_sd(m5stickv, mocker):
+    from krux.pages.device_tests import DeviceTests
+    from krux.input import BUTTON_PAGE
+    from unittest.mock import ANY
+
+    mocker.patch(
+        "uos.statvfs",
+        new=mocker.MagicMock(return_value=[0, 4096, 4096, 0, 1024]),
+    )
+    ctx = create_ctx(mocker, None)
+    tool = DeviceTests(ctx)
+    tool.flash_text = mocker.MagicMock()
+    tool.sd_check()
+    tool.flash_text.assert_has_calls([mocker.call("SD card not detected.", ANY)])
+
+
+def test_sd_check(m5stickv, mocker):
+    from krux.pages.device_tests import DeviceTests
+    from krux.input import BUTTON_ENTER, BUTTON_PAGE
+
+    BTN_SEQUENCE = [
+        BUTTON_ENTER,  # explore files
+    ]
+    mocker.patch(
+        "uos.statvfs",
+        new=mocker.MagicMock(return_value=[0, 4096, 4096, 0, 1024]),
+    )
+    mocker.patch(
+        "os.listdir",
+        new=mocker.MagicMock(return_value=["somefile", "otherfile"]),
+    )
+    ctx = create_ctx(mocker, BTN_SEQUENCE)
+    tool = DeviceTests(ctx)
+    tool.sd_check()
+    ctx.display.draw_hcentered_text.assert_has_calls(
+        [
+            mocker.call(
+                "SD card\n\nSize: 16 MB\n\nUsed: 12 MB\n\nFree: 4 MB",
+                highlight_prefix=":",
+            )
+        ]
+    )
+    assert ctx.input.wait_for_button.call_count == len(BTN_SEQUENCE)
+
+
 def test_printer_test_tool(amigo, mocker, mocked_print_qr):
     """Test that the print tool is called with the correct text"""
     from krux.pages.device_tests import DeviceTests
